v0.2.0 changes:   -Script completely refactored as the previous commit was only a concept test      *Configuration files were created      *Folder (named lib) holds scripts that source functions that can be reused by other scripts   -Better language-package search that work for all packages that are installed and also optionally require them   -Added multiple unofficial repositories to gather binary packages instead of building them all from AUR      *archlinuxcn      *sergej-repo      *archstrike      *archlinuxgr-any      *xyne-any   -User files are created as root and then the owner is changed, instead of creating the files as the user itself   -User is added to a group with GID identical to his UID, instead of adding him to "users"   -Default themes changed:      *Adapta-Nokto-Eta replaces Arc-Grey-Dark as main theme      *Adapta-Nokto replaces Arc-Grey-Dark as window theme      *Papirus-Adapta-Nokto replaces Numix-Circle as icon theme   -ASCII and terminal colors changed to match new Adapta-Nokto theme (in lightblue/cyan instead of white/gray)   -Added script for arranging user desktop after install. It removes itself after running successfully   -First wifi connection is now handled by netctl instead of wpa_supplicant.      *netctl is the default wifi handler until NetworkManager is installed, then NetworkManages is the default      *NetworkManager remains as the default connection manager   -Weather function in terminal (thanks to Orajnam from ManjaroLinux forum!)      *use "weather <city or postalcode>", ex: "weather new york"   -Various files that were copied from darchos-script (in formatting microSD process) are now in this repository, so that repository remains only with essential resources   -Added ALPMP hook to reinstate os-release file on kernel install/upgrade   -Changed default visudo text editor to micro   -Added alias for 'cd..' so it runs 'cd ..'   -Fix:      */tmp is not resized if value is empty in config.txt file   -Removed some packages that were not that useful for regular users from the list of packages to install

+++ res/user-home/tmp.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 
-## Temporary script that will arrange desktop and remove itself
+## Temporary script that will arrange desktop and remove itself after the desktop was successfully arranged
 ## This script should run only for the first time user logged in
 ## Regular users wouldn't even notice it
 
-sleep 10
+sleep 15
 xfdesktop --arrange
-rm ~/.config/autostart/arrange.desktop
-rm ~/.tmp.sh
+if [[ $? -eq 0 ]]; then
+    rm ~/.config/autostart/arrange.desktop
+    rm ~/.tmp.sh
+fi
+xfdesktop --reload
